# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_ingest_api_v1_data_assets_ingest_post": {
        "properties": {
          "declaration": {
            "maxLength": 4096,
            "title": "Declaration",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "project_id": {
            "maxLength": 128,
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "declaration",
          "file"
        ],
        "title": "Body_ingest_api_v1_data_assets_ingest_post",
        "type": "object"
      },
...
      "CreateRequest": {
        "additionalProperties": false,
        "properties": {
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "project_id",
          "spec"
        ],
        "title": "CreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntakeDeclaration": {
        "additionalProperties": false,
        "properties": {
          "license": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "License",
            "type": "string"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          },
          "source": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Source",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 9999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "name",
          "source",
          "license"
        ],
        "title": "IntakeDeclaration",
        "type": "object"
      },
      "JsonValue": {},
      "LocalImport": {
        "additionalProperties": false,
        "properties": {
          "declaration": {
            "$ref": "#/components/schemas/IntakeDeclaration"
          },
          "path": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "project_id": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Project Id",
            "type": "string"
          },
          "source": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "source",
          "path",
          "declaration"
        ],
        "title": "LocalImport",
        "type": "object"
      },
...
      "UpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "spec": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "expected_version",
          "spec"
        ],
        "title": "UpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CoastMAS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/data-assets": {
      "get": {
        "operationId": "list_data_api_v1_data_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/JsonValue"
                    },
                    "type": "object"
                  },
                  "title": "Response List Data Api V1 Data Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Data"
      },
      "post": {
        "operationId": "create_data_api_v1_data_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Create Data Api V1 Data Assets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Data"
      }
    },
    "/api/v1/data-assets/ingest": {
      "post": {
        "operationId": "ingest_api_v1_data_assets_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_api_v1_data_assets_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Ingest Api V1 Data Assets Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/data-assets/ingest-local": {
      "post": {
        "operationId": "ingest_local_api_v1_data_assets_ingest_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Ingest Local Api V1 Data Assets Ingest Local Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Local",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/data-assets/local-files": {
      "get": {
        "operationId": "local_files_api_v1_data_assets_local_files_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Local Files Api V1 Data Assets Local Files Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Local Files",
        "tags": [
          "data"
        ]
      }
    },
...
    "/api/v1/data-assets/{identifier}": {
      "delete": {
        "operationId": "archive_api_v1_data_assets__identifier__delete",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive",
        "tags": [
          "data"
        ]
      },
      "get": {
        "operationId": "get_data_api_v1_data_assets__identifier__get",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Get Data Api V1 Data Assets  Identifier  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Data"
      },
      "put": {
        "operationId": "update_data_api_v1_data_assets__identifier__put",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": {
              "title": "Identifier",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Update Data Api V1 Data Assets  Identifier  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Data"
      }
    },
...
  }
}